src/services/material_bank_parser/requirements.txt
specklepy==2.6.7
Flask==2.0.1
gunicorn==20.1.0
mediatr==1.3.2
python-dotenv